Usuaris, creacio, la compra, i web general

+++ usuaris/views.py
@@ -6,7 +6,7 @@ from django.core.urlresolvers import reverse
 from django.forms import modelform_factory
 from .models import Usuari
 from django.conf import settings
-from jocs.models import Joc
+from jocs.models import Joc, Comprat
 from comandes.models import Comanda, Carret
 from django.db.models import Q
 from django.contrib.auth.models import User
@@ -32,8 +32,11 @@ def crear_usuari(request, perfil_id=None):
             else:
                 password = form.cleaned_data['password']
                 #creem el nou usuari
-                nou_usuari = User.objects.create_user( username = email, email = email, password = password )
+                nou_usuari = User.objects.create_user( username = email, email = email, password = password)
                 nou_carro = Carret.objects.create(usuari = nou_usuari)
+                usuari = Usuari.objects.get( usuari = nou_usuari )
+                usuari.monedes = 100
+                usuari.save()
                 messages.info(request,"Usuari creat correctament")
                 return redirect('jocs:index')
     else:
@@ -65,11 +68,12 @@ def crear_admin(request, perfil_id=None):
             else:
                 password = form.cleaned_data['password']
                 #creem el nou usuari
-                nou_usuari = User.objects.create_user( username = email, email = email, password = password )
+                nou_usuari = User.objects.create_user( username = email, email = email, password = password)
+                nou_carro = Carret.objects.create(usuari = nou_usuari)
                 usuari = Usuari.objects.get( usuari = nou_usuari )
+                usuari.monedes = 100
                 usuari.admin = True
                 usuari.save()
-                nou_carro = Carret.objects.create(usuari = nou_usuari)
                 messages.info(request,"Usuari creat correctament")
                 return redirect('jocs:index')
     else:
@@ -130,65 +134,75 @@ def menu_usuari(request): #Aixo es el carrito de la compra
     return render(request,"menu_usuari.html",ctx)
     
 def biblioteca(request):
-    biblio = Joc.objects.filter(usuari=request.user)
+    usuari = Usuari.objects.get(usuari=request.user)
+    biblio = Comprat.objects.filter(usuari=usuari)
     ctx={"biblio":biblio}
-    return render(request,"menu_usuari.html",ctx)
+    return render(request,"biblioteca.html",ctx)
     
     
 def afegir_al_carritu(request, id_joc):
     carritu = Carret.objects.get(usuari=request.user)
     joc_demanat=Joc.objects.get(id=id_joc)
-    nova_comanda = Comanda.objects.create(carro = carritu, joc = joc_demanat, preu = joc_demanat.preu, quantitat = 1)
-    carritu.preu_total += joc_demanat.preu
+    nova_comanda = Comanda.objects.create(carro = carritu, joc = joc_demanat, preuE = joc_demanat.preuE, preuG = joc_demanat.preuG)
+    carritu.preu_total += joc_demanat.preuE
+    carritu.preuG_total += joc_demanat.preuG
     carritu.save()
     return redirect("usuaris:menu_usuari") 
     
 def eliminar_comanda(request,id_comanda):
     comanda = Comanda.objects.get(id=id_comanda);
     carritu = Carret.objects.get(id=comanda.carro.pk)
-    carritu.preu_total -= comanda.preu
+    carritu.preu_total -= comanda.preuE
+    carritu.preuG_total -= comanda.preuG
+    
+    if carritu.preu_total < 0:
+        carritu.preu_total = 0
+        
+    if carritu.preuG_total < 0:
+            carritu.preuG_total = 0
+    
     carritu.save()
     comanda.delete()
     return redirect('usuaris:menu_usuari')
 
 def modificar_perfil(request):
     usuariForm = modelform_factory(User,fields=("first_name","last_name","email"))
-    profileForm = modelform_factory(Usuari,fields=('direccio','compte'))
     unPerfil = request.user.usuari
     unUsuari = request.user
     
     if request.method == 'POST':
         formUsuari = usuariForm(request.POST, instance = unUsuari)
-        formPerfil = profileForm(request.POST, request.FILES, instance = unPerfil)
+        #formPerfil = profileForm(request.POST, request.FILES, instance = unPerfil)
         formUsuariValid = formUsuari.is_valid()
-        formPerfilValid = formPerfil.is_valid()
+        #formPerfilValid = formPerfil.is_valid()
             
-        if formUsuariValid and formPerfilValid:
+        if formUsuariValid: # and formPerfilValid:
             formUsuari.save()
-            formPerfil.save()
+            #formPerfil.save()
             messages.info(request,"S'han modificat les dades correctament")
             return redirect('jocs:index')
     else:
         formUsuari = usuariForm(instance = unUsuari)
-        formPerfil = profileForm(instance = unPerfil)
+        #formPerfil = profileForm(instance = unPerfil)
         
     for f in formUsuari.fields:
        formUsuari.fields[f].widget.attrs['class'] = 'formulari'
-    for f in formPerfil.fields:
-       formPerfil.fields[f].widget.attrs['class'] = 'formulari'
+    #for f in formPerfil.fields:
+    #  formPerfil.fields[f].widget.attrs['class'] = 'formulari'
    
     formUsuari.fields['first_name'].widget.attrs['placeholder']="Nom"
     formUsuari.fields['last_name'].widget.attrs['placeholder']="Cognoms"
     formUsuari.fields['first_name'].widget.attrs['placeholder']="email"
-    formPerfil.fields['direccio'].widget.attrs['placeholder']="direccio"
-    formPerfil.fields['compte'].widget.attrs['placeholder']="compte"
+    #formPerfil.fields['direccio'].widget.attrs['placeholder']="direccio"
+    #formPerfil.fields['compte'].widget.attrs['placeholder']="compte"
     
-    return render(request, 'modificar_usuari.html', {'formPerfil': formPerfil, 
-                                                     'formUsuari': formUsuari } )
+    return render(request, 'modificar_usuari.html', { 'formUsuari': formUsuari } )
                                                      
                                                      
 from django.core.urlresolvers import reverse
 from django.shortcuts import render
+
+
 from paypal.standard.forms import PayPalPaymentsForm
 
 def view_that_asks_for_money(request, id_carritu):
@@ -243,12 +257,25 @@ def show_me_the_money(sender, **kwargs):
 
 
 def pagat(request, id_carritu):
-    carretbuit=Carret.objects.get(id=id_carritu)
-    comandes=Comanda.objects.filter(carro=carretbuit).delete()
-    carretbuit.preu_total = 0
-    carretbuit.save()
-    return redirect('usuaris:menu_usuari')
-    
+    carretbuit=Carret.objects.get(id=id_carritu) #Carro amb comandes
+    usuari = Usuari.objects.get(usuari=carretbuit.usuari) #Usuari del carro
+    if usuari.monedes >= carretbuit.preuG_total: #Si te les monedes suficients...
+        #agafar els jocs i deixar-los en comprats
+        elsjocs = Comanda.objects.filter(carro=carretbuit)
+        for joc in elsjocs:
+            nou_comprat = Comprat.objects.create( usuari = usuari, joc = joc.joc , completat= False)
+            
+        usuari.monedes = usuari.monedes - carretbuit.preuG_total
+        usuari.save()
+        comandes=Comanda.objects.filter(carro=carretbuit).delete()
+        carretbuit.preu_total = 0
+        carretbuit.preuG_total = 0
+        carretbuit.save()
+        
+        return redirect('usuaris:menu_usuari')
+    else:
+        messages.error(request,"No tens les monedes necesaries")
+        return redirect('usuaris:menu_usuari')
      
-     
+
